Fix BND module name detection for multi-release jars
If a dependency is a multi-release JAR with the module descriptor in
`META-INF/versions/9/module-info.class`, BND is not able to detect its
module name.

We fix the names of the two multi-release JARs in Apache Logging
Services (`log4j-api` and `log4j-core`). The same fix should be applied
to all multi-release external dependencies in children POMs.