Selectively disable reproducibility check on build
We add a `reproducibility-check-enabled` option to the reusable build
that will allow us to turn off the reproducibility check on certain
builds.

While the reproducibility check might detect some problems in a release
build, IMHO it only increases the build time in most builds.

Reproducibility problems are usually rare and are mostly due to
the introduction of additional plugins.
Also note that some kinds of reproducibility problems can only be
detected by running the build on a **different** machine.
We already encountered problems that only depended on the absolute path
of the folder, where the build runs.