Adds `maven-args` option to workflows
We add `maven-args` option to the:

* `build-reusable.yaml` workflow.
* `merge-dependabot-reusable.yaml` workflow.

This parameter sets the
[`MAVEN_ARGS`](https://maven.apache.org/configure.html#maven_args-environment-variable)
environment variable that can be used to pass additional options (e.g. `-Prelease`, `-Dslf4j.version=...`) to the Maven runs.

Note that we don't add an equivalent option for
[`MAVEN_OPTS`](https://maven.apache.org/configure.html#maven_opts-environment-variable),
since the JVM parameters for Maven can be stored in a `.mvn/jvm.config` file.