Build website on the fastest runner only
Site generation logic in `build-reusable.yaml` is currently running on all three runners slowing down builds. When the build is running on a `push` trigger, we can disable site generation entirely, by setting `site-enabled` to `false` and letting `deploy-site-reusable.yaml` do its job. This is no longer possible in PRs, because `deploy-site-reusable.yaml` does not run there.

This PR disables site generation on MacOS and Windows, since we don't expect any differences in the way the website is generated on those platforms.